LayoutTests/imported/w3c/web-platform-tests/css/css-borders/corner-shape/corner-shape-valid-expected.txt


PASS e.style['corner-top-left-shape'] = "round" should set the property value
PASS e.style['corner-top-left-shape'] = "scoop" should set the property value
PASS e.style['corner-top-left-shape'] = "notch" should set the property value
PASS e.style['corner-top-left-shape'] = "bevel" should set the property value
PASS e.style['corner-top-left-shape'] = "squircle" should set the property value
PASS e.style['corner-top-left-shape'] = "square" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-top-left-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-top-right-shape'] = "round" should set the property value
PASS e.style['corner-top-right-shape'] = "scoop" should set the property value
PASS e.style['corner-top-right-shape'] = "notch" should set the property value
PASS e.style['corner-top-right-shape'] = "bevel" should set the property value
PASS e.style['corner-top-right-shape'] = "squircle" should set the property value
PASS e.style['corner-top-right-shape'] = "square" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-top-right-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-bottom-left-shape'] = "round" should set the property value
PASS e.style['corner-bottom-left-shape'] = "scoop" should set the property value
PASS e.style['corner-bottom-left-shape'] = "notch" should set the property value
PASS e.style['corner-bottom-left-shape'] = "bevel" should set the property value
PASS e.style['corner-bottom-left-shape'] = "squircle" should set the property value
PASS e.style['corner-bottom-left-shape'] = "square" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-bottom-left-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-bottom-right-shape'] = "round" should set the property value
PASS e.style['corner-bottom-right-shape'] = "scoop" should set the property value
PASS e.style['corner-bottom-right-shape'] = "notch" should set the property value
PASS e.style['corner-bottom-right-shape'] = "bevel" should set the property value
PASS e.style['corner-bottom-right-shape'] = "squircle" should set the property value
PASS e.style['corner-bottom-right-shape'] = "square" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-bottom-right-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-start-start-shape'] = "round" should set the property value
PASS e.style['corner-start-start-shape'] = "scoop" should set the property value
PASS e.style['corner-start-start-shape'] = "notch" should set the property value
PASS e.style['corner-start-start-shape'] = "bevel" should set the property value
PASS e.style['corner-start-start-shape'] = "squircle" should set the property value
PASS e.style['corner-start-start-shape'] = "square" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-start-start-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-start-end-shape'] = "round" should set the property value
PASS e.style['corner-start-end-shape'] = "scoop" should set the property value
PASS e.style['corner-start-end-shape'] = "notch" should set the property value
PASS e.style['corner-start-end-shape'] = "bevel" should set the property value
PASS e.style['corner-start-end-shape'] = "squircle" should set the property value
PASS e.style['corner-start-end-shape'] = "square" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-start-end-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-end-start-shape'] = "round" should set the property value
PASS e.style['corner-end-start-shape'] = "scoop" should set the property value
PASS e.style['corner-end-start-shape'] = "notch" should set the property value
PASS e.style['corner-end-start-shape'] = "bevel" should set the property value
PASS e.style['corner-end-start-shape'] = "squircle" should set the property value
PASS e.style['corner-end-start-shape'] = "square" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-end-start-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-end-end-shape'] = "round" should set the property value
PASS e.style['corner-end-end-shape'] = "scoop" should set the property value
PASS e.style['corner-end-end-shape'] = "notch" should set the property value
PASS e.style['corner-end-end-shape'] = "bevel" should set the property value
PASS e.style['corner-end-end-shape'] = "squircle" should set the property value
PASS e.style['corner-end-end-shape'] = "square" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(2)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(.5)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(7)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(0.3)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(  0)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(2 )" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(infinity)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(-infinity)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(-0.5)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(-4)" should set the property value
PASS e.style['corner-end-end-shape'] = "superellipse(calc(0.5 * 4))" should set the property value
PASS e.style['corner-shape'] = "round round round round" should set the property value
PASS e.style['corner-shape'] = "round scoop" should set the property value
PASS e.style['corner-shape'] = "round scoop round scoop" should set the property value
PASS e.style['corner-shape'] = "bevel superellipse(2)" should set the property value
PASS e.style['corner-shape'] = "superellipse(0.5) superellipse(3) square" should set the property value
PASS e.style['corner-shape'] = "superellipse(-0.5) superellipse(3) square superellipse(-30)" should set the property value
PASS e.style['corner-shape'] = "superellipse(0.5) superellipse(3) superellipse(1)" should set the property value
PASS e.style['corner-shape'] = "bevel superellipse(2) squircle     round" should set the property value
PASS e.style['corner-shape'] = "superellipse(0.5) superellipse(3) superellipse(1) superellipse(infinity)" should set the property value
FAIL e.style['corner-top-shape'] = "scoop" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-top-shape'] = "scoop" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-top-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-top-shape'] = "bevel notch" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-top-shape'] = "bevel notch" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-top-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-top-shape'] = "superellipse(-2) squircle" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-top-shape'] = "superellipse(-2) squircle" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-top-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-top-shape'] = "superellipse(3)" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-top-shape'] = "superellipse(3)" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-top-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-right-shape'] = "scoop" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-right-shape'] = "scoop" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-right-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-right-shape'] = "bevel notch" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-right-shape'] = "bevel notch" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-right-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-right-shape'] = "superellipse(-2) squircle" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-right-shape'] = "superellipse(-2) squircle" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-right-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-right-shape'] = "superellipse(3)" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-right-shape'] = "superellipse(3)" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-right-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-bottom-shape'] = "scoop" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-bottom-shape'] = "scoop" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-bottom-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-bottom-shape'] = "bevel notch" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-bottom-shape'] = "bevel notch" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-bottom-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-bottom-shape'] = "superellipse(-2) squircle" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-bottom-shape'] = "superellipse(-2) squircle" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-bottom-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-bottom-shape'] = "superellipse(3)" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-bottom-shape'] = "superellipse(3)" should set corner-bottom-right-shape assert_equals: corner-bottom-right-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-bottom-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-left-shape'] = "scoop" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-left-shape'] = "scoop" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-left-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-left-shape'] = "bevel notch" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-left-shape'] = "bevel notch" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-left-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-left-shape'] = "superellipse(-2) squircle" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-left-shape'] = "superellipse(-2) squircle" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-left-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-left-shape'] = "superellipse(3)" should set corner-bottom-left-shape assert_equals: corner-bottom-left-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-left-shape'] = "superellipse(3)" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-left-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-start-shape'] = "scoop" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-inline-start-shape'] = "scoop" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-inline-start-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-start-shape'] = "bevel notch" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-inline-start-shape'] = "bevel notch" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-inline-start-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-start-shape'] = "superellipse(-2) squircle" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-inline-start-shape'] = "superellipse(-2) squircle" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-inline-start-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-start-shape'] = "superellipse(3)" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-inline-start-shape'] = "superellipse(3)" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-inline-start-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-end-shape'] = "scoop" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-inline-end-shape'] = "scoop" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-inline-end-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-end-shape'] = "bevel notch" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-inline-end-shape'] = "bevel notch" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-inline-end-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-end-shape'] = "superellipse(-2) squircle" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-inline-end-shape'] = "superellipse(-2) squircle" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-inline-end-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-inline-end-shape'] = "superellipse(3)" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-inline-end-shape'] = "superellipse(3)" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-inline-end-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-start-shape'] = "scoop" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-block-start-shape'] = "scoop" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-block-start-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-start-shape'] = "bevel notch" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-block-start-shape'] = "bevel notch" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-block-start-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-start-shape'] = "superellipse(-2) squircle" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-block-start-shape'] = "superellipse(-2) squircle" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-block-start-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-start-shape'] = "superellipse(3)" should set corner-start-end-shape assert_equals: corner-start-end-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-block-start-shape'] = "superellipse(3)" should set corner-start-start-shape assert_equals: corner-start-start-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-block-start-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-end-shape'] = "scoop" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-block-end-shape'] = "scoop" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "scoop" but got ""
FAIL e.style['corner-block-end-shape'] = "scoop" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-end-shape'] = "bevel notch" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "notch" but got ""
FAIL e.style['corner-block-end-shape'] = "bevel notch" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-block-end-shape'] = "bevel notch" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-end-shape'] = "superellipse(-2) squircle" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "squircle" but got ""
FAIL e.style['corner-block-end-shape'] = "superellipse(-2) squircle" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "superellipse(-2)" but got ""
FAIL e.style['corner-block-end-shape'] = "superellipse(-2) squircle" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-block-end-shape'] = "superellipse(3)" should set corner-end-end-shape assert_equals: corner-end-end-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-block-end-shape'] = "superellipse(3)" should set corner-end-start-shape assert_equals: corner-end-start-shape should be canonical expected "superellipse(3)" but got ""
FAIL e.style['corner-block-end-shape'] = "superellipse(3)" should not set unrelated longhands assert_true: expected true got false
